app/requirements.txt
protobuf==3.20.*
grpcio
grpcio-tools
grpcio-reflection